refactor: BaseViewController 구현하여 공통부분 추상화
