+++ OpenMarket/OpenMarket/Controller/MainViewControllerUnderiOS14.swift
@@ -7,11 +7,7 @@
 
 import UIKit
 
-class MainViewControllerUnderiOS14: UIViewController {
-    enum Section {
-        case main
-    }
-    
+class MainViewControllerUnderiOS14: BaseViewController {
     private let dataProvider = DataProvider()
     private var products: [Product] = [] {
         didSet {
@@ -20,96 +16,19 @@ class MainViewControllerUnderiOS14: UIViewController {
             }
         }
     }
-    
-    private var collectionView: UICollectionView?
-    private var baseView = BaseView()
-    private var listLayout: UICollectionViewLayout?
-    private var gridLayout: UICollectionViewLayout?
+    private var isFirstSnapshot = true
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        setUpNavigationItem()
-        applyListLayout()
-        applyGridLayout()
-        configureHierarchy(collectionViewLayout: listLayout ?? UICollectionViewLayout())
         setUpCollectionView()
-        
+        view.backgroundColor = .systemBackground
         dataProvider.fetchData() { products in
             self.products.append(contentsOf: products)
         }
     }
-}
-
-extension MainViewControllerUnderiOS14 {
-    private func setUpCollectionView() {
-        collectionView?.register(ProductListCell.self, forCellWithReuseIdentifier: "ProductListCell")
-        collectionView?.register(ProductGridCell.self, forCellWithReuseIdentifier: "ProductGridCell")
-        collectionView?.dataSource = self
-        collectionView?.delegate = self
-        collectionView?.backgroundColor = .systemBackground
-    }
     
-    private func setUpNavigationItem() {
-        setUpSegmentation()
-        navigationItem.titleView = baseView.segmentedControl
-        navigationItem.rightBarButtonItem = UIBarButtonItem(title: "+", style: .plain, target: self, action: #selector(touchUpRegisterProduct))
-    }
-    
-    private func setUpSegmentation() {
-        baseView.segmentedControl.setWidth(view.bounds.width * 0.18 , forSegmentAt: 0)
-        baseView.segmentedControl.setWidth(view.bounds.width * 0.18, forSegmentAt: 1)
-        baseView.segmentedControl.addTarget(self, action: #selector(switchCollectionViewLayout), for: .valueChanged)
-    }
-    
-    @objc private func touchUpRegisterProduct() {
-        let registerProductView = UpdateProductViewController()
-        let navigationController = UINavigationController(rootViewController: registerProductView)
-        navigationController.modalTransitionStyle = .coverVertical
-        navigationController.modalPresentationStyle = .fullScreen
-        present(navigationController, animated: true)
-    }
-    
-    @objc private func switchCollectionViewLayout() {
-        switch baseView.segmentedControl.selectedSegmentIndex {
-        case 0:
-            guard let listLayout = listLayout else {
-                return
-            }
-            collectionView?.setCollectionViewLayout(listLayout, animated: false)
-            collectionView?.reloadData()
-        case 1:
-            guard let gridLayout = gridLayout else {
-                return
-            }
-            collectionView?.setCollectionViewLayout(gridLayout, animated: false)
-            collectionView?.reloadData()
-        default:
-            break
-        }
-    }
-}
-
-extension MainViewControllerUnderiOS14 {
-    private func configureHierarchy(collectionViewLayout: UICollectionViewLayout) {
-        collectionView = UICollectionView(frame: .zero, collectionViewLayout: listLayout ?? collectionViewLayout)
-        view.addSubview(collectionView ?? UICollectionView())
-        layoutCollectionView()
-    }
-    
-    private func applyGridLayout() {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .fractionalHeight(1.0))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .fractionalHeight(0.3))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitem: item, count: 2)
-        group.interItemSpacing = .fixed(10)
-        let section = NSCollectionLayoutSection(group: group)
-        section.contentInsets = NSDirectionalEdgeInsets(top: 5, leading: 5, bottom: 5, trailing: 5)
-        section.interGroupSpacing = 10
-        let layout = UICollectionViewCompositionalLayout(section: section)
-        gridLayout = layout
-    }
-    
-    private func applyListLayout() {
+    // MARK: override function (non @objc)
+    override func applyListLayout() {
         let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0),
                                               heightDimension: .fractionalHeight(1.0))
         let item = NSCollectionLayoutItem(layoutSize: itemSize)
@@ -120,22 +39,25 @@ extension MainViewControllerUnderiOS14 {
         let layout = UICollectionViewCompositionalLayout(section: section)
         listLayout = layout
     }
+}
+
+// MARK: CollectionView Setting
+extension MainViewControllerUnderiOS14 {
+    private func setUpCollectionView() {
+        collectionView?.register(ProductListCell.self, forCellWithReuseIdentifier: "ProductListCell")
+        collectionView?.register(ProductGridCell.self, forCellWithReuseIdentifier: "ProductGridCell")
+        collectionView?.dataSource = self
+        collectionView?.delegate = self
+        collectionView?.backgroundColor = .systemBackground
+    }
     
-    private func layoutCollectionView() {
-        guard let collectionView = collectionView else {
-            return
-        }
-        collectionView.translatesAutoresizingMaskIntoConstraints = false
-        
-        NSLayoutConstraint.activate([
-            collectionView.topAnchor.constraint(equalTo: view.topAnchor),
-            collectionView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            collectionView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            collectionView.trailingAnchor.constraint(equalTo: view.trailingAnchor)
-        ])
+    @objc override func switchCollectionViewLayout() {
+        super.switchCollectionViewLayout()
+        collectionView?.reloadData()
     }
 }
 
+// MARK: UICollectionViewDataSource
 extension MainViewControllerUnderiOS14: UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return products.count
@@ -169,6 +91,7 @@ extension MainViewControllerUnderiOS14: UICollectionViewDataSource {
     }
 }
 
+// MARK: UICollectionViewDelegate
 extension MainViewControllerUnderiOS14: UICollectionViewDelegate {
     func collectionView( _ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
         guard let _ = products[safe: indexPath.row + 1] else {
@@ -176,6 +99,30 @@ extension MainViewControllerUnderiOS14: UICollectionViewDelegate {
                 self.products.append(contentsOf: products)
             }
             return
+        }
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        guard let product = products[safe: indexPath.row] else { return }
+        
+        let registerProductView = UpdateProductViewController()
+        let navigationController = UINavigationController(rootViewController: registerProductView)
+        navigationController.modalTransitionStyle = .coverVertical
+        navigationController.modalPresentationStyle = .fullScreen
+        
+        HTTPManager().loadData(targetURL: .productDetail(productNumber: product.identifier)) { productDetail in
+            switch productDetail {
+            case .success(let productDetail):
+                guard let decodedData = try? JSONDecoder().decode(ProductDetail.self, from: productDetail) else {
+                    return
+                }
+                registerProductView.initialize(product: decodedData)
+                DispatchQueue.main.async { [self] in
+                    present(navigationController, animated: true)
+                }
+            case .failure(_):
+                return
+            }
         }
     }
 }
